Add commit-msg, pre-push, and pre-rebase hooks
This adds several git hook scripts that are meant to be added to a given
git repository under the $GIT_DIR/hooks directory.  This is meant to
serve as an automatic linter of git commit messages and to ask the user
questions about whether they're following the process regarding usage of
fixup/squash commits, rebasing, etc.

The commit-msg hook is used to enforce the git commit message format for
the title, separator and line length of the commit message body.  It
also checks for the proper use of fixup and squash commits.  If the
commit message isn't formatted properly, if a fixup or squash commit is
not properly written, or if the title contains information that's not
informative, then the commit will not be saved.

The pre-push hook is used to ask the user questions about whether
they've informed the rest of the team about their rebase if they're
doing a force push and asks if the user has checked the commit messages
and associated diffs.  If the user answers yes to both questions, only
then will the push be allowed to go through.

The pre-rebase hook asks the user whether they're doing a rebase or git
commit --amend to address a review comment and tells the user to use a
fixup or squash commit instead.  It also reminds the user to check the
diff between the upstream branch and their local branch after rebase to
verify whether any changes were introduced by the rebase.  It will
suggest to run a git reset --hard to reset the branch to the state
before the rebase if unintended changes were introduced.  It also
reminds the user to check the log to verify the commit messages and
their associated diffs look correct before pushing.  If the user answers
yes to the question about rebasing/amending to address a comment, then
it rejects the attempt to rebse.