Add post-rewrite and post-commit hooks
The post-rewrite hook is used to remind the user to use "fixup! " or
"squash! " commits instead of the amend or rebase they just did.  The
output gives instructions on how to reset the branch to the state before
the amend or rebase and basic instructions on how to make a "fixup! " or
"squash! " commit.

The post-commit hook is used to remind the user to put a link in a
Github PR comment in response to a comment made on the PR that the
commit is meant to address.  This makes it easier for the reviewer to
find the commit that makes the changes requested and makes it easier for
the person making the commit to find the sha1 of the commit they just
made.

Neither of these hooks will affect the ability to do a rebase, amend, or
conventional commit.  They will only print the information described
above.