Add fetch detection to pre-rebase hook
When rebasing a branch, we want to make sure that the latest changes in
the master branch are incorporated into the branch that's under review.
This checks whether the sha1 of the remote tracking branch for master
and the sha1 of the corresponding branch in the remote repository match.
If they don't, then that means the user should run a git fetch origin or
git pull origin to update the remote tracking branch (and the
corresponding local branch if running git pull).

If the sha1 values don't match, the rebase attempt is rejected.